+++ test/excel/excel_reader_test.dart
@@ -6,7 +6,7 @@ import 'package:test/test.dart';
 main() {
   group("ExcelReader", () {
     test("Reads a sheet correctly", () {
-      List<int> randomNumbers = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
+      List<double> randomNumbers = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
 
       Excel excel = ExcelFileBuilder(randomNumbers: randomNumbers).buildExcel();
       var excelReader = ExcelReader(book: excel);
